feat(collections): stabilize iterable input of takeLastWhile
