+++ collections/take_last_while_test.ts
@@ -55,3 +55,64 @@ Deno.test("takeLastWhile() returns the same array when all elements match the pr
 
   assertEquals(actual, [1, 2, 3, 4]);
 });
+
+Deno.test("takeLastWhile() handles generator", () => {
+  function* gen() {
+    yield 1;
+    yield 2;
+    yield 3;
+    yield 4;
+    yield 5;
+    yield 6;
+  }
+  const actual = takeLastWhile(gen(), (i) => i !== 4);
+  assertEquals(actual, [5, 6]);
+});
+
+Deno.test("takeLastWhile() returns empty array when the last generator element does not match the predicate", () => {
+  function* gen() {
+    yield 1;
+    yield 2;
+    yield 3;
+    yield 4;
+  }
+
+  const actual = takeLastWhile(gen(), (i) => i !== 4);
+  assertEquals(actual, []);
+});
+
+Deno.test("takeLastWhile() returns the same array when all elements match the predicate", () => {
+  function* gen(): Generator<number> {
+    yield 1;
+    yield 2;
+    yield 3;
+    yield 4;
+  }
+  const actual = takeLastWhile(gen(), (i) => i !== 400);
+  assertEquals(actual, [1, 2, 3, 4]);
+});
+
+Deno.test("takeLastWhile() empty generator returns empty array", () => {
+  function* gen(): Generator<number> {}
+  const actual = takeLastWhile(gen(), (i) => i > 4);
+  assertEquals(actual, []);
+});
+
+Deno.test("takeLastWhile() gets from last matching element from an array", () => {
+  const arr = [1, 2, 3, 4, 5, 6];
+  const actual = takeLastWhile(arr, (i) => i !== 2 && i !== 4);
+  assertEquals(actual, [5, 6]);
+});
+
+Deno.test("takeLastWhile() gets from last matching element from a generator", () => {
+  function* gen(): Generator<number> {
+    yield 1;
+    yield 2;
+    yield 3;
+    yield 4;
+    yield 5;
+    yield 6;
+  }
+  const actual = takeLastWhile(gen(), (i) => i !== 2 && i !== 4);
+  assertEquals(actual, [5, 6]);
+});
